requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-mdeditor==0.1.18
importlib-metadata==4.8.1
Markdown==3.3.4
paramiko==2.8.0
pycparser==2.20
pyDes==2.0.1
Pygments==2.10.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tornado==6.1
torndb==0.3
typing-extensions==3.10.0.2
zipp==3.6.0
